AD_gen/requirements.txt
tweepy
psycopg2-binary
python-dotenv
selenium 
bs4
transformers
serpapi
google-search-results
torch 
newspaper3k
psycopg2
openai==0.28.0
webdriver-manager
django-environ